+++ packages/grid/_modules_/grid/hooks/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './useLogger';
 export * from './useScrollFn';
-export * from './useThemeProps';
