+++ packages/grid/_modules_/grid/utils/index.ts
@@ -6,3 +6,4 @@ export * from './keyboardUtils';
 export * from './mergeUtils';
 export * from './paramsUtils';
 export * from './getLocalization';
+export * from './material-ui-utils';
